moved header buttons into components
