+++ Lexicon/client/src/components/headerMobile/HeaderMobileContent.js
@@ -1,0 +1,10 @@
+import React from 'react'  
+// Slide in from right and take up almost/all of the screen width
+
+const HeaderMobileContent = () => (
+    <nav>
+        NAV
+    </nav>
+)
+
+export default HeaderMobileContent
