+++ Lexicon/client/src/styles/button.css
@@ -15,7 +15,15 @@
 .btn__no-style {
   border: none;
   background-color: transparent;
+  padding: 0;
+}
+
+.btn__no-style:hover {
   cursor: pointer;
+  color: var(--darkGray);
+}
+
+.btn__hamburger {
   padding: 0 0.5em 0 0;
 }
